+++ changelog.md
@@ -1,6 +1,7 @@
 ## Forestry 2.1.1
 - Fix IIndividualHandlerItem methods returning individuals for empty stacks that were decremented
 - Added method in IAlleleManager to get IChromosome instances by registry name
+- Fix butterfly humidity tolerance having the same range as temperature tolerance
 
 ## Forestry 2.1.0
 - Fixed typo that cut off black order on the right side of the Portable Analyzer GUI